--Akhilesh

+++ src/main/java/com/ats/adminpanel/model/ggreports/GGReportGrpByFrId.java
@@ -12,11 +12,11 @@ public class GGReportGrpByFrId {
 	
 	private int isGrn;
 	
-	float aprGrandTotal;
-	float totalAmt;
+	Float aprGrandTotal;
+	Float totalAmt;
 	String frName;
 	int reqQty;
-	int aprQty ;
+	Float aprQty ;
 	public int getFrId() {
 		return frId;
 	}
@@ -29,16 +29,16 @@ public class GGReportGrpByFrId {
 	public void setIsGrn(int isGrn) {
 		this.isGrn = isGrn;
 	}
-	public float getAprGrandTotal() {
+	public Float getAprGrandTotal() {
 		return aprGrandTotal;
 	}
-	public void setAprGrandTotal(float aprGrandTotal) {
+	public void setAprGrandTotal(Float aprGrandTotal) {
 		this.aprGrandTotal = aprGrandTotal;
 	}
-	public float getTotalAmt() {
+	public Float getTotalAmt() {
 		return totalAmt;
 	}
-	public void setTotalAmt(float totalAmt) {
+	public void setTotalAmt(Float totalAmt) {
 		this.totalAmt = totalAmt;
 	}
 	public String getFrName() {
@@ -53,10 +53,10 @@ public class GGReportGrpByFrId {
 	public void setReqQty(int reqQty) {
 		this.reqQty = reqQty;
 	}
-	public int getAprQty() {
+	public Float getAprQty() {
 		return aprQty;
 	}
-	public void setAprQty(int aprQty) {
+	public void setAprQty(Float aprQty) {
 		this.aprQty = aprQty;
 	}
 	@Override
@@ -64,5 +64,6 @@ public class GGReportGrpByFrId {
 		return "GGReportGrpByFrId [frId=" + frId + ", isGrn=" + isGrn + ", aprGrandTotal=" + aprGrandTotal
 				+ ", totalAmt=" + totalAmt + ", frName=" + frName + ", reqQty=" + reqQty + ", aprQty=" + aprQty + "]";
 	}
+	
 
 }
